Added DNDSnapView - the base listener from which views (such as buttons) are drop.
Updates
DNDPager -
 void render() views are loaded when called.
 void generateSnapGrid() loads DNDSnapViews with onDragListeners
 void generateButton() loads a button inside the layout.

MainActivity - integrated the DNDPager with Relative layout to render the views.

activity_main.xml - remove background color in relative layout, since it will the DNDPager class overrides the layout's background.
                    use DNDPager.setBackground() to set the preferred background instead.

Note: file that are not mentioned but still committed are not yet official and may be removed in later updates.